pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "raggen"
version = "0.1.2"
authors = [
  { name="Alex Karev", email="karev.as99@gmail.com" },
]
description = "RAG dataset generator"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
  "langchain[openai]==0.3.18",
  "langchain_text_splitters==0.3.6",
  "pydantic==2.10.6",
  "tenacity==9.0.0",
  "tiktoken==0.8.0",
  "jinja2==3.1.5",
  "tqdm==4.67.1",
  "pandas==2.2.3",
  "marker-pdf==1.3.5",
  "html2text==2024.2.26",
  "mammoth==1.9.0"
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/alex-karev/birbtalk"
Issues = "https://github.com/alex-karev/birbtalk/issues"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format='%(asctime)s - %(levelname)s - %(message)s'
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
